Refatorar integrações do Supabase em módulos dedicados
